Changes to the class structure of the entity package.
Should not affect any code using Player, Wall and Bullet.

I realised that the MovableRectangularEntity class contained a lot of movement functions that are not specific to rectangles.
To fix this I need to have both Rectangular entity and movable entity to extend entity. This then means that classes which
are both rectangular and movable must inherit both. Multiple inheritance is not possible in Java therefore the MovableRectangularEntity
class now serves to combine them via composition. This is not an ideal solution but I can't come up with a better one.

I plan to reduce the number of constructors when I have created a system that will read player info from a file.